Fixed preferences selection view with an overlaying box layout and updated newsources API query
