+++ app/src/main/java/com/example/palexis3/newssum/composable/articles/NewsApiArticleDetailsScreen.kt
@@ -1,4 +1,4 @@
-package com.example.palexis3.newssum.composable
+package com.example.palexis3.newssum.composable.articles
 
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
